Busy polling on pipes in Windows's emulation of select
The previous implementation of the select emulation on pipes would
often wait for 16ms when no data is immediately available (or a better
resolution if another process happens to request such a resolution).

This can lead to bad reactivity of programs that rely on select.  This
commit attempts to fix this by switching to an busy polling for about
64ms (but making sure to yield to other threads/processes).  After
this initial period, we go back to loop with 16ms timeouts.  Thus, the
time "wasted" is at most 1/4 of the total wait time.  The downside is
that the process can eat 100% of a one CPU core if data comes fast
enough.

An alternative might to do a blocking read on the pipe (in a thread),
with a zero buffer size.